Fix build for GCC. Cast enum to int when using with %d format specifier
TBR: robertphillips@google.com

No-Tree-Checks: true
No-Try: true

Bug: skia:8962

+++ src/gpu/ops/GrTransferFromOp.h
@@ -34,7 +34,7 @@ public:
         string.appendf(
                 "bufferID:: %d offset: %zu, color type: %d\n"
                 "srcRect: [ L: %d, T: %d, R: %d, B: %d ]\n",
-                fDstBuffer->uniqueID().asUInt(), fDstOffset, fDstColorType, fSrcRect.fLeft,
+                fDstBuffer->uniqueID().asUInt(), fDstOffset, (int)fDstColorType, fSrcRect.fLeft,
                 fSrcRect.fTop, fSrcRect.fRight, fSrcRect.fBottom);
         return string;
     }
